requirements.txt
django>=4.0.0,<4.1.0
djangorestframework
psycopg2
pillow
django-cors-headers
pyyaml
django-rest-swagger